requirements.txt
torch
torchvision
timm
faiss-cpu
matplotlib
uvicorn
fastapi
timm
python-multypart
